refactor: move also file writing into the package
