+++ src/opossum_lib/opossum_file_model/constants.py
