+++ tests/test_cli.py
@@ -11,12 +11,12 @@ from _pytest.logging import LogCaptureFixture
 from click.testing import CliRunner, Result
 
 from opossum_lib.cli import generate
-from opossum_lib.input_formats.opossum.constants import (
+from opossum_lib.opossum_file_model.constants import (
     INPUT_JSON_NAME,
     OUTPUT_JSON_NAME,
 )
-from opossum_lib.input_formats.opossum.opossum_file_writer import OpossumFileWriter
 from opossum_lib.opossum_file_model.opossum_file import OpossumPackage
+from opossum_lib.opossum_file_model.opossum_file_writer import OpossumFileWriter
 from tests.test_setup.opossum_file_faker_setup import OpossumFileFaker
 
 test_data_path = Path(__file__).resolve().parent / "data"
